make column header links bold
Closes: CORE-846

the "Students" header is bold because that is what the styleguide calls 
for. but the rest of the column headers are font-weight "normal" because
instUI links resets them back to that. but kdub says links in <th> 
should still be bold

Test plan:

* turn on user/course search feature flag
* go to /accounts/x
* all the column headers should be bold
* go to /accounts/x/users
* all the column headers should be bold

Change-Id: I2d2970bb68fb67db3a8b1058b8084948ad3259c6
Reviewed-on: https://gerrit.instructure.com/137480
Tested-by: Jenkins
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>