use uploadFile to upload via url
closes RECNVS-331

when the upload is provided a url instead of a file, the preflight
response will include a progress object. once the upload requests are
complete (immediately after preflight for non-instfs, after the post to
the upload_url for instfs), we still need to wait for either canvas or
instfs to finish cloning the provided url, so we poll the progress
object. once that's complete, the completed progress object's results
will have the attachment information.

test-plan:
- have g/142111 checked out in your inst-fs repo
- have inst-fs enabled
- in a course, register an LTI integration that can return File Items
  (jacob, ahuff, or nathanm can clarify)
- add an assignment with file upload submission type
- as a student, submit the assignment using a file selected file through
  the LTI tool
- the UI flow for submission should work
- the submission attachment should end up on inst-fs
- the submission should be downloadable

Change-Id: I3d4d4b0e0aad852a8452a64b5526e41e179ff5b0
Reviewed-on: https://gerrit.instructure.com/142642
QA-Review: Collin Parrish <cparrish@instructure.com>
Reviewed-by: Andrew Huff <ahuff@instructure.com>
Tested-by: Jenkins
Product-Review: Jacob Fugal <jacob@instructure.com>