spec: restrict jest filename matching
By default, jest uses the following matchers:

  '**/__tests__/**/*.js?(x)'
  '**/?(*.)(spec|test).js?(x)'

It looks for .js and .jsx files inside of __tests__ folders, as well as
any files with a suffix of .test or .spec (e.g. Component.test.js or
Component.spec.js). It will also find files called test.js or spec.js.

This means that ANY file in the __tests__ folders is considered to be a
test file and will fail if it does not contain any specs. The prohibits
the use of supporting helper files and/or other such code to aid in
testing.

The change in this commit restricts the matcher to only files contained
within __tests__ paths with names ending in spec.js or test.js
(e.g., thing.test.js, spec.js)

Change-Id: Ied771b9e8dd5ad8fd24a9de70f1edb9a272ed3c1
Reviewed-on: https://gerrit.instructure.com/142657
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Product-Review: Jeremy Neander <jneander@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>