Automated RTL css fixes for gradebooks
refs: CORE-1226

I split this out from the javascript changes for slick grid because
the test plan for these changes can be much simpler.
the net result of this change is that absolutely nothing
should change in any of our existing css output. this just makes it so
all of these places where we said:
margin/padding/float/border/border-radus: (left or right), it does the
opposite in RTL so it flows right.

While the review on this can be completely automated, it does need to
be submitted at the same time as the one that changes the slickgrid
javascript. So when you want to manually QA this, check this and that
out at the same time and QA them together.

test plan:
* before you check this out, compile all the css in app.
* copy the contens of public/dist/brandable_css to a folder
* check this commit out
* run brandable_css again and diff the output of that folder with
  the old folder.
* there should be absolutely no changes

* verify that for the "_rtl" variants, the right and left values are
  swapped whenever they need to be.

With both this and the next commit that does all the RTL js changes
for slickgrid checked out:
* turn on the RTL feature flag and browse around, all the grading pages
  should be RTL and all of the margins, padding, spacing, arrow
  icons, etc. should be flipped and make sense for an RTL layout.

Change-Id: I896fe581f1ec91383387d34055509bea5b5c41b9
Reviewed-on: https://gerrit.instructure.com/145412
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
Tested-by: Jenkins
Reviewed-by: Jeremy Neander <jneander@instructure.com>
QA-Review: Tucker McKnight <tmcknight@instructure.com>
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>