Don't show warning on ungraded discussions.
If we have a graded discussion assigned to only some sections
of a course, and mark it ungraded, no longer show a warning
that it's not assigned to everyone.

Apologies for not knowing how to write a spec around this.

Fixes COMMS-1014

Test Plan:
* Have a course with multiple sections.
* Have a discussion assigned using the "overrides" to only
  some of those sections.
* Now, from the discussions edit page, mark that discussion
  as ungraded.
* There should no longer be a warning shown.

Change-Id: I9c8fb4812ffb0c7f18eb89d9d8c55bdfc8100886
Reviewed-on: https://gerrit.instructure.com/146073
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Aaron Kc Hsu <ahsu@instructure.com>
Product-Review: Venk Natarajan <vnatarajan@instructure.com>