ensure screenreader_gradebook can access ‘direction’
fixes: CORE-1328

This bundle includes it’s own copy of a bunch of the bootstrap css.
It should not do that in the first place since it is already included
In the common css bundle but it is, and some subtle styles depend on
It being loaded again in this bundle. So if it is going to do that,
It needs to have access to our ‘direction’ rtl sass helpers

Test plan:
* smoke test on screenreader gradebook page to make sure that styles
  still look the same as they did before

Change-Id: I96108b0f98dcddb1211aa21a4c0ed63125dda534
Reviewed-on: https://gerrit.instructure.com/147813
Tested-by: Jenkins
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>