mark ‘allow_rtl’ as beta not ‘development’
fixes: CORE-1326

If something is marked as ‘development’ it cannot be turned on in
production. If it is marked as ‘beta’ it will just show a ‘beta’
Label but doesn’t affect where it can be turned on.

I didn’t realize this and we made a change to change this feature from
state: ‘hidden’, to state: 'allowed' so that accounts could turn this on
but without this change they still won’t be able to. So we need to
warmfix this.


Test plan:
The “Allow RTL users to see RTL interface” should show a “beta”
Tag next to it but it should be turn-on-able by any account admin (not
Just site admins)

Change-Id: Ib1ce574f7c2548593b330bf726d9b9c24214be34
Reviewed-on: https://gerrit.instructure.com/147951
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>