UpdateItemTray closes when it should
fixes ADMIN-1055

note: skipping the selenium spec "student planner Create To Do Sidebar
edits a To Do" because it refuses to pass. I firmly believe it's due
to selenium running against stale planner code, so skipping to get
this change moving forward until we
figure out what's going on in the build

It seems that the newer version of Tray calls the onExited callback
later than it used to. It was in our handler for onExited that we
cleared out the updateTodoItem. Because that was happening too late, the
tray never closed.  We get the previous order of execution by attaching
to the onExit event, which fires as soon as the Tray is toggled closed.

test plan:
  - click on the + button to create a new item
  - fill it in and click Save
  > expect the tray to close and focus to be on the new item
  - click on the new item
  - click on the X close button in the tray
  > expect focus to return to the item you clicked
  - click on the new item
  > expect the tray to open
  - make a change and click Save
  > expect the tray to close and the item to reflect your
    change in the  planner and focus to be on the item
  - click on the item
  > expect the tray to open
  - click on Delete and confirm
  > expect the tray to close and for the item to be gone and
    focus to be on the prior item.

Change-Id: I4cc094898f879e2dc2c6f9572b87288a9a9e5ebc
Reviewed-on: https://gerrit.instructure.com/150557
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Luke Kingsley <lkingsley@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>