fix dashboard list view edit tray while scrolling
In the dashboard list view, if the edit tray was open and you scrolled
the main view such that the new activity button would disappear or
reappear, it would wipe out the changes you'd made in the edit tray.

This also removes the scrolling key handlers because according to the
a11y peeps, this shouldn't have been a thing anyway, and we don't want
the arrow keys conflicting with the arrow key handling of the instui
components in the tray.

fixes ADMIN-1094

test plan:
- load the planner
- click the + button and add a title
- scroll around and make the "new activity" button disappear and
  reappear
  > The title should remain the same and not be cleared
- clicking on a to do item in the main view should still change the data
  in the edit tray to reflect the clicked item.

Change-Id: I613f6ad853a68cc67d3e18abdac3af5f698a326d
Reviewed-on: https://gerrit.instructure.com/152036
Tested-by: Jenkins
Reviewed-by: Mysti Sadler <mysti@instructure.com>
QA-Review: Luke Kingsley <lkingsley@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>