fix "Not calculated.." message for assignments
The message was showing every time for assignments in assignment
groups because the default group_weight of a group is zero. This
checks that the assignment group is supposed to be weighted before
showing a message that the weight of the group is zero. The fix is
cosmetic; grades were actually being calculated correctly.

fixes GRADE-1218

Test Plan
- Create an assignment group.
- Add an assignment to the assignment group.
- Navigate to New Gradebook.
- Open the Submission Tray of that assignment for a student.
- Verify that the Submissions Tray does not have a message stating
that the assignment is not calculated for final grade.
- Optionally edit the assignment groups to have weights and set the
assignment group to have a weight of zero.
- Verify that the Submission Tray does have a message stating that
the assignment is not calculated for final grade.

Change-Id: Ie4eec4174860408f3ad22361c2dfd00804934fee
Reviewed-on: https://gerrit.instructure.com/153114
Reviewed-by: Nicholas Pitrak <npitrak@instructure.com>
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
Tested-by: Jenkins
QA-Review: James Butters <jbutters@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>