Remove specious outcome extra credit alerts
"Extra credit not possible on outcome" alerts were incorrectly appearing
when navigating from a student to another student when both students
were graded using a rubric with an outcome-based criterion and the new
student had fewer points on the criterion than the old student.  This
happened regardless of whether the relevant feature flag was enabled,
and was the result of comparing data in the rubric summary and the
editable form.  Only perform the relevant comparison when the grader
has entered new scores, and not when we are switching between students.

fixes GRADE-1198

Test plan:

Setup:
- Have a course with some students
- Define an outcome for the course
- Set up a rubric with a criterion mapped to the outcome
- As a grader, open SpeedGrader for the assignment and assess some
  students using the rubric

With the feature flag "Outcome Extra Credit" enabled:
- Attempt to enter a score higher than the outcome-based criterion's
  maximum allowed points and save
- Saving the assessment should succeed and keep the score you entered

With the flag disabled:
- Attempt to enter a score higher than the outcome-based criterion's
  maximum allowed points and save
- Saving the assessment should produce an "extra credit not possible"
  alert and force the score to the maximum allowed points

Testing that specious alerts don't show up:
- Assess students such that one student has a score on the criterion
  that is higher than a student immediately next to them in the list
  of students (but both are at/below the maximum points)
- Navigating back and forth between the two students should not produce
  any "extra credit not possible" alerts
- Similarly, editing one student's assessment and switching students
  without saving your changes should not produce errors
  - Note that currently we do not save any changes made when navigating
    this way (this is pre-existing behavior)

Change-Id: I412f743624268850e757781cbc80641d48227c77
Reviewed-on: https://gerrit.instructure.com/155287
Reviewed-by: Gary Mei <gmei@instructure.com>
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
Tested-by: Jenkins
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>