Make the permissions with no search matches show correctly
fixes COMMS-1225

Test Plan
* Make sure that new permissions is on
* navigate to the new permissions page
* Use the search bar and find a search with no results
* Notice the cornerstone permissions label does not shrink
anymore
* Note when i originally started this ticket i thought it
was a different problem, but that problem could not be
reproduced. It might be nice to check if there are any
other related styling problems and see if we need to this
ticket if necessary.

Change-Id: Idd196c89aa45c1f087ff2805428adafd40bd42ca
Reviewed-on: https://gerrit.instructure.com/155687
Reviewed-by: Venk Natarajan <vnatarajan@instructure.com>
Tested-by: Jenkins
QA-Review: Sara Chadwick <schadwick@instructure.com>
Product-Review: Aaron Kc Hsu <ahsu@instructure.com>