fix width of checkboxes on discussions edit page
The issue was the checkboxe label were much longer than their text. The
trick I used to fix it was to wrap the labels in a div so they could still
act like blocks, but make the label itself an inline-block so the
label would only be as long as its text.

fixes ADMIN-1213

test plan:
- Open the discussions add/edit page
- You shouldn't be able to click on the checkboxes to the right of their
label text.
- Click on the "Add student to-do" checkbox, and then open the date
picker. Click to the whitespace to the right of the checkboxes to close
the date picker. It shouldn't interact with any of the checkboxes.

Change-Id: I9057da3f5eb07b70d8952e4a280964b8834305c2
Reviewed-on: https://gerrit.instructure.com/158235
Tested-by: Jenkins
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>