Fix focus outline from view calendar link
Canvas is moving to a 1px outline on focused items.
In the normal canvas theme, the view calendar link gets
an outline and underline on focus.
With underlined links enabled, the link gets an outline and the
underline disappears on focus.
In high-contrast it's the same, but the outline is darker.

closes ADMIN-1260

test plan:
  - load the card dashboard
  - press TAB a thousand times until the Show Calendar link gets focus
  > expect the view calendar link to be underlined and have a 1px outline
  - enable underlined links from the user's settings
  > expect the view calendar link to have the outline, but the underline
    that was there before focus disappears
  - enable high contrast
  > expect the view calendar link to behave the same as before, but the
    outline is darker
  - one of these times, press ENTER while the link has focus
  > expect to be taken to the calendar

Change-Id: I95f0668da77ab630e1bc23e8a3a0d7a30c2a824a
Reviewed-on: https://gerrit.instructure.com/159898
Reviewed-by: Mysti Sadler <mysti@instructure.com>
Reviewed-by: Dan Minkevitch <dan@instructure.com>
Tested-by: Jenkins
Product-Review: Dan Sasaki <dsasaki@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>