fix dismissing opportunities in list view
I added assignment_id to the PlannerOverride json so that we can easily
link a plannable to its associated assignment, if it has one. Before we
were just assuming that the override was on the assignment, but now
we've standardized to have the override on the assignment's associated
object, if it has one. Since opportunities loads assignments, we needed
a way to easily navigate from the override back to the assignment so the
opportunities menu could update the proper data.

fixes ADMIN-1214

test plan:
- Create a student with missing assignments of several types:
  - regular assignments
  - discussions
  - quizzes
- The student should be able to dismiss all of these types of
assignments from the opportunities menu
- Undismiss the opportunities in the rails console so they show up again
  for example:
  PlannerOverride.
    where(user: <student>, dismissed: true).
    update_all(dismissed: false)
- This will repopulate the opportunities drop down. The student should
  still be able to dismiss all the opportunities.
- Dismiss opportunities very quickly, as described in the Jira ticket.
  There should be no errors.

Change-Id: Ia96dd9d8c970e330d07b5a185a4f8b1195a7ee43
Reviewed-on: https://gerrit.instructure.com/161248
Reviewed-by: Mysti Sadler <mysti@instructure.com>
Tested-by: Jenkins
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Mysti Sadler <mysti@instructure.com>