remove informStudentsOfOverdueSubmissions, nothing uses it
It looks like back in g/134730 and g/134729 we added this to only
Show the late and missing pills to new gradebook users. But then
In g/153679 we stopped caring about if it was new or old gradebook
And showed them all the time. So nothing uses this attribute anymore.

Test plan:
* nothing should change since nothing uses this attribute

Change-Id: Id3543dc34534809f96c6b7538599bea4a2152eb0
Reviewed-on: https://gerrit.instructure.com/161906
Tested-by: Jenkins
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>