Fix dismissing to-dos from course ToDoSidebar
fixes ADMIN-1383

the planner api urls in the code were relative, which failed when
the user was on a course page. Changed them to absolute urls.

don't render the "Show All" button in the sidebar if there are no items
in the sidebar.

added a selenium specs to test that dismissing a to-do from the sidebar
causes it to show up as complete in planner

test plan:
  - student with course-related to-dos and modules as the course home
    page
  - as the student, go to the course
  > expect the to-do(s) to show up in the To Do sidebar
  - disconnect from the network
  - click on the 'x' to dismiss an item
  > expect a flash alert telling you it failed
  - reconnect to the network
  - click the 'x' to dismiss an item
  > expect it to disapper from the list
  - go to the planner
  > expect the dismissed item to show up as complete

  - repeat, but with a course whose home page is the activity stream,
    because that's subtly different.

Change-Id: Ieb0c59d0a2d510be3d2b559bcb6d0658260b4588
Reviewed-on: https://gerrit.instructure.com/163543
Tested-by: Jenkins
Reviewed-by: Mysti Sadler <mysti@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>