Record audit events when rubric set
Record an appropriate event when a rubric is associated with, or
dissociated from, an auditable assignment.

closes GRADE-1488

Test plan:

(The easiest way to check that events are being created properly is
probably to run AnonymousOrModerationEvent.last (or similar) in a Rails
console. See below for the payload format of each event described here.)

- Create an anonymous (or otherwise auditable) assignment
- Create at least one grading rubric
- On the assignment view screen, attach a rubric to the assignment
  - This should create an event of type rubric_created
- Without deleting that rubric, attach a different rubric
  - This should create an event of type rubric_updated
- Remove the rubric from the assignment
  - This should create an event of type rubric_deleted
- Attach a rubric to the assignment, then go to the course rubrics page
  (from the "Manage Rubrics" link in the rubric selection dialog)
  and delete the rubric
  - This should create an event of type rubric_deleted

- For a non-auditable assignment (i.e., neither anonymous nor moderated)
  make sure that attaching or detaching a rubric as above does NOT
  create any events

Format of the payload for each event type:
- rubric_created: {id: <added rubric ID>}
- rubric_updated: {id: [<old rubric ID>, <new rubric ID>]}
- rubric_deleted: {id: <removed rubric ID>}

(Note that these should be the IDs of the *rubrics* themselves, not of
the association objects mapping the rubric to the assignment.)

Change-Id: Id467cfbb2482734874e303284e7a9dab15ecdb7d
Reviewed-on: https://gerrit.instructure.com/163827
Tested-by: Jenkins
Reviewed-by: Gary Mei <gmei@instructure.com>
Reviewed-by: Jeremy Neander <jneander@instructure.com>
QA-Review: James Butters <jbutters@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>