remove unused slickgrid files
These are all the js files from slickgrid that are not used in any of
our webpack bundles. So technically, we don’t need them.

Gradebook people, what would you prefer? To get rid of them so we don’t 
have files sitting around in our repo that are unused or keep them
so whatever is in /vendor/slickgrid is an accurate representation of
what was copypasta’ed into our codebase?

It is up to you, I am just making this commit because I was auditing 
All our other unused stuff and so you are aware. I’m good either way. 
If you want me to submit this, I’ll fix up the commitmessage

Also if we are going to delete these I’m sure there are .png, .gif and 
.css files in public/javascripts/vendor/slickgrid that are also unused
that could be deleted if we wanted to

Test plan:
* all the automated builds should pass
* the gradebook, gradezilla and jobs pages should work

Change-Id: Ia0be29b959c7b264c828531b5a0d6e839e9a58a3
Reviewed-on: https://gerrit.instructure.com/167745
Tested-by: Jenkins
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Keith Garner <kgarner@instructure.com>
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>