prefer display name over filename, but encode it
fixes SAS-418, SAS-338

inst-fs will read the segment of the path following the UUID and attempt
to use it as the filename in the content-disposition header when it
serves the file. prefer using the display name over the filename for
that, so that when if the file is served as a download (vs. inline) it
is saved according to the display name on the user's machine. but we
need to make sure to URI encode it before putting it in the path, in
case of any special characters in the display name.

test-plan:
 - have inst-fs enabled
 - upload a file
 - rename the file
 - download the file; should save according to the new name, not the
   original name
 - rename the file to have special characters in it (spaces,
   URI-specific symbols like "?" or "/", non-UTF-8 characters, etc.)
 - download the file; should save according to the new name, but the URL
   downloaded from (i.e. check the inst-fs URL redirected to) should
   have those characters escaped

Change-Id: Ic6fb3cbafffb51791605c7a6ba2e29117213576e
Reviewed-on: https://gerrit.instructure.com/174739
Tested-by: Jenkins
Reviewed-by: Michael Jasper <mjasper@instructure.com>
QA-Review: Michael Jasper <mjasper@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>